requirements.txt
absl-py==1.4.0
astunparse==1.6.3
attrs==22.2.0
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1666764759924/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools==5.3.0
certifi==2022.12.7
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1671179893800/work
charset-normalizer==3.0.1
ConfigArgParse==1.5.3
contourpy @ file:///Users/runner/miniforge3/conda-bld/contourpy_1673633760692/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography-split_1677821374538/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
dataclasses==0.6
flatbuffers==23.3.3
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1666827370036/work
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio @ file:///Users/runner/miniforge3/conda-bld/grpcio_1653138968159/work
h5py @ file:///Users/runner/miniforge3/conda-bld/h5py_1637964553571/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1663332044897/work
keras @ file:///home/conda/feedstock_root/build_artifacts/keras_1669020828024/work/keras-2.11.0-py2.py3-none-any.whl
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1666805765141/work
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib @ file:///Users/runner/miniforge3/conda-bld/matplotlib-suite_1676406471798/work
munkres==1.1.4
networkx==3.0
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1673482170163/work
pandas==1.5.3
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1675487373585/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1677883256757/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1677600624926/work
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1672659226110/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
pyroaring==0.3.8
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1673864280276/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1673863902341/work
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/scikit-learn_1674599166167/work
scikit-mine==0.0.11
scipy==1.10.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sortedcontainers==2.4.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-macos==2.11.0
tensorflow-metal==0.7.1
termcolor==2.2.0
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1643647933166/work
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1666788685513/work
typing_extensions==4.5.0
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1667239979860/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1673452138552/work
Werkzeug==2.2.3
wget==3.2
wrapt==1.15.0
xattr==0.10.1
